Remove FunctionMockUtil class and move functions to Main
